try fix accuraty

+++ JumpKnock/app/src/main/java/ch/band/jumpknock/game/UiNotifier.java
@@ -5,8 +5,8 @@ import java.util.List;
 public interface UiNotifier {
 	void addPlatform(Platform platform);
 	void removePlatform(Platform platform);
-	void UpdateGame(List<Platform> platforms, Player player, int reachedHeight);
-	void updateUi(int height);
+	void UpdateGame(List<Platform> platforms, Player player, float reachedHeight);
+	void updateUi(float height);
 	void gameOver(int height);
 	void playerCollidedWith(Platform platform);
 	float getSmartPhoneRotation();
